sync

+++ train_patch.py
@@ -103,8 +103,8 @@ ghs, shs, gh_sample, sh_sample = [[None] * args.nr_gpu for i in range(4)]
 
 if args.global_conditional:
     latent_dim = 2
-    gh_init = tf.placeholder(tf.int32, shape=(args.init_batch_size, latent_dim))
-    ghs = [tf.placeholder(tf.int32, shape=(args.batch_size, latent_dim)) for i in range(args.nr_gpu)]
+    gh_init = tf.placeholder(tf.float32, shape=(args.init_batch_size, latent_dim))
+    ghs = [tf.placeholder(tf.float32, shape=(args.batch_size, latent_dim)) for i in range(args.nr_gpu)]
 if args.spatial_conditional:
     latent_shape = obs_shape[0], obs_shape[1], obs_shape[2]+1 ##
     sh_init = tf.placeholder(tf.float32, shape=(args.init_batch_size,) + latent_shape)
